requirements.txt
crewai>=0.95.0
crewai-tools>=0.17.0
python-dotenv>=1.0.1
tavily-python>=0.5.0
groq>=0.11.0
litellm>=1.50.0
streamlit>=1.35.0
